+++ resources/[dg]/dg-jobs/typescript/src/server/modules/postop/controller.postop.ts
@@ -1,5 +1,11 @@
 import { Events, RPC } from '@dgx/server';
-import { finishJobForGroup, startDropoff, finishDropoff, startJobForGroup } from './service.postop';
+import {
+  finishJobForGroup,
+  startDropoff,
+  finishDropoff,
+  startJobForGroup,
+  skipCurrentLocation,
+} from './service.postop';
 
 Events.onNet('jobs:postop:signIn', (src: number, jobType: PostOP.JobType) => {
   startJobForGroup(src, jobType);
@@ -11,3 +17,7 @@ Events.onNet('jobs:postop:finish', (src: number, netId: number) => {
 
 RPC.register('jobs:postop:startDropoff', startDropoff);
 Events.onNet('jobs:postop:finishDropoff', finishDropoff);
+
+Events.onNet('jobs:postop:skipLocation', plyId => {
+  skipCurrentLocation(plyId);
+});