+++ resources/[dg]/dg-jobs/typescript/src/server/modules/postop/service.postop.ts
@@ -260,3 +260,30 @@ export const finishDropoff = (plyId: number, dropoffId: number, success: boolean
     }
   });
 };
+
+export const skipCurrentLocation = (plyId: number) => {
+  const group = getGroupByServerId(plyId);
+  if (!group) return;
+  const active = activeGroups.get(group.id);
+  if (!active) return;
+
+  active.locationSequence.shift();
+  active.locationsDone++;
+  active.dropoffsBusy.clear();
+  active.dropoffsDone.clear();
+
+  // Check if all dropoffs are done for group
+  let finishedAllLocations = false;
+  if (active.locationSequence[0] !== undefined) {
+    active.targetLocation = getRandomTargetLocation(active.type, active.locationSequence[0]);
+  } else {
+    finishedAllLocations = true;
+  }
+
+  const phoneNotificationData = buildPhoneNotificationData(active);
+  group.members.forEach(m => {
+    if (!m.serverId) return;
+    Phone.updateNotification(m.serverId, 'postop_amount_tracker', phoneNotificationData);
+    Events.emitNet('jobs:postop:setLocation', m.serverId, finishedAllLocations ? null : active.targetLocation);
+  });
+};